WIP: arm64: dts: meson: khadas-vim3: change LED behaviours on VIM3/3L
VIM3/3L have two LEDs (red/white) located 1mm apart on the board. In the
current configuration both are powered on creating a pink visual status
indication. The boards are also marketed as HTPC devices where heartbeat
pulses are distracting and undesirable.

This change removes the heartbeat and sets the default state so that only
the white LED is on while the board is powered.

Signed-off-by: Christian Hewitt <christianshewitt@gmail.com>