Revert "of/fdt: Make sure no-map does not remove already reserved regions"
U-Boot on Hardkernel ODROID-N2 adds fdt "memreserve" entries alongside
the reserved-memory nodes present in upstream Linux. This causes
reserving memory with the no-map flag to fail:
[    0.000000] OF: fdt: Reserved memory: failed to reserve memory for node 'secmon@5000000': base 0x0000000005000000, size 3 MiB

This seems to cause kernel crashes such as:
[129988.642342] SError Interrupt on CPU4, code 0xbf000000 -- SError

Restore the bahavior of Linux kernel 5.9 and older and don't fail
updating the mapping with no-map.

This reverts commit 8a5a75e5e9e55de1cef5d83ca3589cb4899193ef.

Signed-off-by: Stefan Agner <stefan@agner.ch>